Switch from double to single precision

+++ src/main/scala/apsu/demo/rocks/components/geometry/AngularVelocity.scala
@@ -5,4 +5,4 @@ package apsu.demo.rocks.components.geometry
  *
  * @author david
  */
-case class AngularVelocity(theta: Double)
+case class AngularVelocity(theta: Float)
